backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain
langchain-cohere
langchain-groq
cohere
pdfplumber==0.10.2
python-docx==0.8.11
pytesseract==0.3.10
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
tiktoken==0.5.2
pillow==10.0.1
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
